toolkit/requirements.txt
# Simplingua Toolkit Requirements
psycopg2-binary>=2.9.0
numpy>=1.24.0
